Create basic follow plus instant alerts button groundworks
At this point this is just a copy of the follow button but with some of the functionality removed.

This extra functionality is not needed at this point for the instant alerts variant.

To avoid conflicts and bugs I think its easier to have the instant alert version as a new file with only the functionality it needs.

Should the follow plsu instant alerts test be successful we will need to revisit and review all the logic.