Position the preference modal relative to its trigger
This is a basic solution that allows the preference modal to be rendered in the correct position on the page based on the button that triggered it.

We have built this in a way that was quick and easy and that we know works given the constraints of the test it will be used in. If this test is successful
and this feature should be used elsewhere much of this will need rethinking to make it more generic and reusable.

Co-authored-by: ytcleon <leon.yam@ft.com>