Remove myFT add + instant alert test
In Sept 2023 we added this as a test to see if we could increase the awareness
of the instant alert feature.

This allowed users to add topics to instant alerts at the same time as adding the topic to
their myFT as a "follow".

This test is now being removed because some of the code for it was done in a quick and hacky way
for the test. As no decision has been made on the success of the test we wish to tidy up by
removing the test.